Project description.
This is more like a draft. I written what I want, but I don't know if I
will be able to implement everything. I may also see new usages, or
decide that some of my ideas are not so good and remove them. That's my
first public project, we will see!