Added a `handle` script to launch our code.
I added a link to have the README with a `.coffee.md` extension, and a
little script that will call our literate CoffeeScript program. That
script is like a shell script, except it is written in CoffeeScript (I
hope that if someone is able to launch that script, whatever the
platform be, the script will be able to launch our code (I'm looking at
you windows. You're not developper friendly, will you not complain about
the PATH?)).