Added dependencies.
In fact, I wrote a big chunk of code. I'll be committing in small
chunks, but they have an overall coherence. It's always hard to
bootstrap the project with small commits. Once the project is more
advanced, we can commit less code at a time, but at the time of
experiments, I tend to write a lot of code _just to see if it works_,
and because I don't know if it will be a good code, I don't commit it.
That may be an error. I need to work on that.