Revisiting the source organization of the projects.
And with that we will have a simpler code, less gotchas, less bugs!

I plan now to have the entry point of our server in
`./src/server.coffee.md` and `./src/client.coffee.md` instead of
`./server.coffee.md` and `./client.coffee.md`. That is because we don't
want to mess with the require paths:

With the first approach, we had to require './src/module' in
coffeescript, but './lib/module' in javascript, and do the translation,
either manually (and code would not seem logical as we require it in an
other location than where we write it), or automatically (and how do we
know that a path must or must not be translated?).