Ahah! Removed Table of Content
It does not look good in github ;-)

Seriously, that may be practical when reading, but it will be a pain in
the ass when writing code, as we need to update the table of content,
that will not be next to our code. Also, maintaining the link references
would have not been nice. And github-specific.

I'll learn to think before commiting things that I will uncommit right
after!

(and should I learn to use that "history rewritting" I've heard of? What
should I do, transparency or clean history?)